requirements.txt
streamlit>=1.32.0
openai>=1.14.0
opencv-python>=4.9.0
Pillow>=10.2.0
python-dotenv>=1.0.1
torch>=2.2.1
torchvision>=0.17.1
diffusers>=0.27.2
transformers>=4.39.1
accelerate>=0.28.0
safetensors>=0.4.2